+++ hal/rtl8814a/usb/rtl8814au_xmit.c
@@ -25,8 +25,8 @@ s32	rtl8814au_init_xmit_priv(_adapter *padapter)
 
 #ifdef PLATFORM_LINUX
 	tasklet_init(&pxmitpriv->xmit_tasklet,
-	     (void(*)(unsigned long))rtl8814au_xmit_tasklet,
-	     (unsigned long)padapter);
+	     (void(*)(uintptr_t))rtl8814au_xmit_tasklet,
+	     (uintptr_t)padapter);
 #endif
 #ifdef CONFIG_TX_EARLY_MODE
 	pHalData->bEarlyModeEnable = padapter->registrypriv.early_mode;
